pyproject.toml
[project]
name = "fast-healthchecks"
version = "0.2.4"
description = "FastHealthchecks"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Vladislav Shepilov", email = "shepilov.v@protonmail.com"},
]
maintainers = [
    {name = "Vladislav Shepilov", email = "shepilov.v@protonmail.com"},
]
keywords = [
    "healthcheck", "library",
    "fastapi", "starlette", "faststream", "litestar",
    "asyncpg", "psycopg", "redis", "aio-pika", "httpx", "aiokafka", "motor",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = "<4.0.0,>=3.10.0"
dependencies = []

[tool.setuptools.package-data]
fast_healthchecks = ["py.typed"]

[tool.setuptools]
packages = [
    "fast_healthchecks",
    "fast_healthchecks.checks",
    "fast_healthchecks.checks.postgresql",
    "fast_healthchecks.integrations",
]

[project.urls]
homepage = "https://github.com/shepilov-vladislav/fast-healthchecks"
documentation = "https://shepilov-vladislav.github.io/fast-healthchecks/"
source = "https://github.com/shepilov-vladislav/fast-healthchecks"
tracker = "https://github.com/shepilov-vladislav/fast-healthchecks/issues"

[project.optional-dependencies]
asyncpg = ["asyncpg>=0.31.0,<1.0.0"]
psycopg = ["psycopg>=3.3.2,<4.0.0"]
redis = ["redis>=7.1.1,<8.0.0"]
aio-pika = ["aio-pika>=9.5.8,<10.0.0"]
httpx = ["httpx>=0.28.1,<1.0.0"]
aiokafka = ["aiokafka>=0.13.0,<1.0.0"]
motor = ["motor>=3.7.1,<4.0.0"]
fastapi = ["fastapi[standard]>=0.129.0,<1.0.0"]
faststream = ["faststream>=0.6.6,<1.0.0"]
litestar = ["litestar>=2.20.0,<3.0.0"]
msgspec = ["msgspec>=0.20.0,<1.0.0"]
opensearch = ["opensearch-py[async]>=3.1.0,<4.0.0"]

[dependency-groups]
dev = [
    "pip-audit>=2.7.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-deadfixtures>=3.1.0",
    "python-dotenv>=1.2.1",
    "pytest-vcr>=1.0.2",
    "pytest-xdist>=3.8.0",
    "greenlet>=3.3.1",
    "pre-commit>=4.5.1",
    "ruff>=0.15.0",
    "ty>=0.0.16",
    "deptry>=0.24.0",
]
docs = [
    "mkdocs-include-markdown-plugin>=7.2.1",
    "mkdocs>=1.6.1,<2.0.0",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.3",
    "pymdown-extensions>=10.20.1",
]

[tool.deptry]
experimental_namespace_package = true
pep621_dev_dependency_groups = ["dev", "docs"]

[tool.deptry.per_rule_ignores]
DEP002 = ["msgspec"]
DEP003 = []
DEP004 = ["dotenv"]

[tool.uv]
default-groups = []

[tool.ruff]
line-length = 120
# Assume Python 3.10 to be compatible with typing.
target-version = "py310"
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".ruff_cache",
    ".tox",
]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = [
    "E501",  # line too long, handled by ruff-format
    "CPY001",  # Missing copyright notice at top of file
    "TC001",  # typing-only imports; can break runtime (Scope, HTTPRouteHandler)
    "TC002",
    "TC003",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"!fast_healthchecks/**.py" = ["D", "S", "SLF"]
"tests/**/*.py" = ["ARG001", "N802", "N806"]

[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
minversion = "8.0"
addopts = "-ra -q"
filterwarnings = [
    "error",
    "error:::fast_healthchecks",
    "ignore:Failing to pass a value to the 'type_params' parameter of 'typing.ForwardRef._evaluate':DeprecationWarning",
    "ignore:Import of deprecated import 'litestar.contrib.attrs.AttrsSchemaPlugin'. Deprecated in litestar 2.13.0. This import will be removed in 3.0. importing AttrsSchemaPlugin from 'litestar.contrib.attrs' is deprecated, please import it from 'litestar.plugins.attrs' instead:DeprecationWarning",
    "ignore::DeprecationWarning:aiohttp.connector",
]
markers = """
    integration: mark a test as an integration test
    unit: mark a test as a unit test
    imports: mark a test as an imports test
"""

[tool.coverage.run]
source = ["fast_healthchecks"]

[tool.coverage.report]
precision = 1
include = ["fast_healthchecks/*"]
omit = [
    "tests/*",
]
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",
    # Don't check obviously not implemented
    "raise NotImplementedError",
    # We don't really care what happens if fail
    "except ImportError:",
    # Don't check for typing-only code
    "if TYPE_CHECKING:",
    # Don't check for code that only runs itself
    "if __name__ == .__main__.:",
]
show_missing = true

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = false
version_files = [
    "pyproject.toml:version",
    "fast_healthchecks/__init__.py:__version__",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
